sdk/requirements.txt
pydantic>=1.8.0
pyyaml>=5.4.0
jsonschema>=3.2.0
